Handle timestamp and time types in Specific Datum Reader
When reading data encoded with one of the various
'timestamp'-logical types, if the field it is being
marshaled into is not an int64 but a time.Time, then this
will panic on unmarshal.  Detect based on the presence of
the 'logicalType' field (and a bad target `reflect.Kind`) to
call into a special function which can handle mapping of
logical types to (in this case) `time.Time` values.

In some ways it would be nicer if marshaling could be
customizable eg via methods, but then you couldn't easily
add it for standard library types like `time.Time` anyway.