Add helpers for matching schemas
The avro spec defines ways in which schemas can be compared
for equality.  It starts by defining a canonical
representation which ignores logical types, docs, etc and
then says to take the Rabin crc64 value.  The spec gives
psuedocode at
https://avro.apache.org/docs/1.8.2/spec.html#schema_fingerprints
for this, and as far as I can tell this is almost exactly
the same as the go implementation of CRC64.